requirements.txt
django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2


